refactor: Restructured codebase
